Restructure files based on feedback
