+++ packages/react-dom/src/fire/ReactFireBatching.js
@@ -9,7 +9,7 @@
 import {
   needsStateRestore,
   restoreStateIfNeeded,
-} from './controlled/ReactFireControlledState';
+} from './host-components/controlled/ReactFireControlledState';
 
 // Used as a way to call batchedUpdates when we don't have a reference to
 // the renderer. Such as when we're dispatching events or if third party
