+++ packages/react-dom/src/fire/host-components/controlled/ReactFireInput.js
@@ -10,13 +10,13 @@
 import invariant from 'shared/invariant';
 import warning from 'shared/warning';
 
-import {getFiberPropsFromDomNodeInstance} from '../ReactFireInternal';
+import {getFiberPropsFromDomNodeInstance} from '../../ReactFireMaps';
 import {getCurrentFiberOwnerNameInDevOrNull} from 'react-reconciler/src/ReactCurrentFiber';
 import {controlledValuePropTypes} from './ReactFirePropTypes';
-import {getToStringValue, toString} from '../ReactFireUtils';
-import type {ToStringValue} from '../ReactFireUtils';
+import {getToStringValue, toString} from '../../ReactFireUtils';
+import type {ToStringValue} from '../../ReactFireUtils';
 import {updateValueIfChanged} from './ReactFireValueTracking';
-import {setValueForProperty} from '../ReactFireComponentProperties';
+import {setHostComponentAttribute} from '../ReactFireHostComponentAttributes';
 
 type InputWithWrapperState = HTMLInputElement & {
   _wrapperState: {
@@ -115,7 +115,7 @@ export function updateChecked(element: Element, props: Object) {
   const domNode = ((element: any): InputWithWrapperState);
   const checked = props.checked;
   if (checked != null) {
-    setValueForProperty(domNode, 'checked', checked, false);
+    setHostComponentAttribute(domNode, 'checked', checked, false);
   }
 }
